requirements.txt
django==1.10.5
djangorestframework==3.8.2
Pillow==5.1.0
wechatpy==1.7.0
easy-thumbnails==2.5
django-jet==1.0.7
django-ckeditor==5.5.0
PyMySQL==0.9.3
python-dateutil==*.*
redis==*.*
django-redis==4.7
qrcode==6.0


